Fixup NARA mapping
 - Extension methods for NARA are lambda-ized and updated to use current
DSL features.
 - `&make_relation` is updated to address a more complex state of affairs
than is originally communicated in the spreadsheet; the
relation/collection hierarchy is more malleable than is described. This
is documented in the comments, but needs to be updated in the
spreadsheet mapping.
 - `&make_rights` is refactored.
 - The `collection` mapping is updated to address the collection hierarchy
as described above.
 - The `date` mapping is updated to handle `coverageDates` with begin and
end date nodes.

 - `description` is updated to correctly match `scopeAndContentNote` as
described in the mapping (values appear directly, rather than nested in
`generalNote.note`
 - `#field('termName')` calls are moved into their respective
`ChildDeclaration` blocks. This avoids splitting values in the case that
more than one `termName` is given. As far as I know, this doesn't
actually happen in production data, but it's slightly safer with the
adjustment.
 - Removes XML formatted comments. These comments are duplicated from the
mapping spreadsheet; including them here spreads maintenance. They are
also somewhat misleading, since the values retrieved are JSON.